+++ src/test/java/calculator/service/CalculatorTest.java
@@ -32,8 +32,8 @@ public class CalculatorTest {
 
     @Test
     void shouldReturnListWithNumbers(){
-        List<Double> expected = List.of(5d,4d,2d,4d);
-        List<Double> actual = calculator.getNumbers("5+4-2+4");
+        List<Integer> expected = List.of(5,4,2,4);
+        List<Integer> actual = calculator.getNumbers("5+4-2+4");
         Assertions.assertEquals(expected, actual);
     }
 }
